requirements.txt
keras==2.13.1
keras==2.12.0
numpy==1.23.5
opencv_contrib_python==4.8.0.74
opencv_python==4.8.0.76
Pillow==9.3.0
Pillow==10.0.0
streamlit==1.24.0
tensorflow==2.12.0
tensorflow_intel==2.13.0

